Add local trigger for ParentReady. (#404)
A lot of code (try_notar, set_timeouts) are only triggered when the
first slot of the window gets ParentReady event.

However, a validator can join the cluster at any slot. If a validator
joins in the middle of a window, currently ParentReady doesn't get
triggered for the first slot, then the timers for later slots in the
window aren't set, and it can't vote Notarize on the slots in the window
either.

This is especially a problem during cluster standstill, if enough
validators are stuck in the state, then the cluster may never recover
from a standstill.

When a validator sees that a later slot in the window is finalized, it
has that block and all its ancestors, mark the first slot of the window
ParentReady, so that we can start voting on rest of the slots in this
window.