parent_ready: choose parent with most skips (#427)
#### Problem
When there are multiple parent readys available to produce our block off
of we should pick the smallest by slot # parent, as that corresponds to
the fork with the most skips.

This is important during duplicate block attacks, as blocks with
multiple notarize-fallback certificates will eventually get a skip and
switching forks in replay incurs a cost which eats into the skip timer.

#### Summary of Changes
Choose the smallest parent ready when selecting the parent for block
production


See
https://anza-xyz.slack.com/archives/C09CD6HK18A/p1756367501898419?thread_ts=1756289889.565439&cid=C09CD6HK18A
for initial discussion