[votor] Fix Fallback certificate build in votor (#432)
#### Problem
There seems to be incorrect logic in the `aggregate` function in
`vote_certificate_builder`. Currently, independent of the type of the
certificate, we always set the signer rank into the first bitmap. The
correct logic should be that if there is a second vote type message that
is aggregated, then the second bitmap bits should be set.

#### Summary of Changes
I fixed the logic and added tests. I also realized that there are no
unit tests for actually verifying BLS signatures on vote or certificate
messages, so I added the tests. On the special case of certificates that
use base3 encodings, the current `solana-bls-signatures` interface does
not provide a function to verify aggregate signatures on different
messages. This is in the works
(https://github.com/anza-xyz/solana-sdk/pull/300). I added a todo to
complete the tests once `solana-bls-signatures` v0.3.0 is out.

Fixes #
<!-- OPTIONAL: Feature Gate Issue: # -->
<!-- Don't forget to add the "feature-gate" label -->