[bls-sigverifier] Cache vote messages (#445)
#### Summary of Changes
BLS optimistic batch verification of votes is much more efficient if we
assume that the signatures are all based on the same message.

I added stats that record the number of distinct messages in a single
batch. I also updated the verification logic so that if there is only
one distinct message for the batch, I use the `par_verify_aggregate`
function instead of the `par_verify_distinct` function.

In the process, I addressed a todo: cache serialized votes so that we
don't serialize identical votes every time.

After we analyze the number of distinct messages in our cluster, we can
add tailored optimizations for the specific number of messages.

Fixes #
<!-- OPTIONAL: Feature Gate Issue: # -->
<!-- Don't forget to add the "feature-gate" label -->

---------

Co-authored-by: Wen <113942165+wen-coding@users.noreply.github.com>