[core] Re-organize batches for bls-sigverify bench (#466)
#### Summary of Changes

I used the logs in the current alpenglow test cluster re-organize the
batches for testing in the bls-sigverify bench logic.

I first attempted to hard-code some of the vote and certificate batches
in the benches, but it was hard to set up the test because the vote and
certificate messages do not contain the actual validator keypairs. So I
updated the code so that the bench logic freshly generates vote and
certificate messages that model the batch of votes and certificate
messages seen commonly in the logs.