feat: actually dump alpenglow metrics (#485)
Closes: #408 

In the earlier PR, we collected the metrics but did not actually dump
them. In this PR, we are also dumping the metrics.

The `Datapoint` API is quite restrictive requiring `&'static str` for
various arguments forcing a lot of code duplication. On my long list of
issues to explore is to see if we can make the `Datapoint` API a bit
more friendlier to use.