feat: `ConsensusMetrics`: improvements to when the slots start (#496)
For `ConsensusMetrics`: previously we were starting slots in the
following manner:
- the first slot in the window started `DELTA_TIMEOUT` time after parent
ready was seen.
- the remaining slots started when the skip timer for the previous slots
fired

In the alpenglow cluster, we observed that this is too pessimistic and
in vast majority of cases, the votes and the block hash arrived before
the slots were started.

This PR improves when the slots start.  In particular:
- the first slot in the window starts when we see the parent ready
- the remaining slots start when either the skip timer for the previous
slot fires; a block for the previous slot was seen; or a finalization
certificate for the previous slot was seen.

Finally as `ConsensusMetrics` is owned by a single struct now, it
updates it so that it is no longer stored inside a lock.