Refactor BLS sigverifier out of the sigverify stage (#475)
#### Problem

Currently, there is redundant serialization logic in the
`verify_batches` and `send_packets` of the bls sigverifier.

#### Summary of Changes
I removed the redundant serialization logic. It wasn't possible to
remove the redundant logic without tweaking the `SigVerifier` trait, so
I added an associated `VerifiedOutput` type to the trait. This change
made computing the stats a little more complicated, so I also added
`VerifyPacketBatchesStats` to simplify the organization a bit.

From the perspective of the original ed25519 sigverifier, the changes
should be the changes should be pretty minimal.

Fixes #
<!-- OPTIONAL: Feature Gate Issue: # -->
<!-- Don't forget to add the "feature-gate" label -->