bug: fix for metrics collection (#526)
#### Problem


https://anza-xyz.slack.com/archives/C08D0FL5Q4T/p1760033928310479?thread_ts=1760017436.833329&cid=C08D0FL5Q4T

Originally standstill was reported as an integer and in
https://github.com/anza-xyz/alpenglow/pull/501 it was changed to be a
boolean. This causing conflicts in influxdb and causing problems with
metrics collection.

#### Summary of Changes

Renames the standstill variable in metrics. Also adds a comment about
when we can rename it back because the original integer data should be
dropped due to the retention policy.