migration: Add genesis vote tracking (#532)
Split from #502 

#### Problem
As part of the migration we need a new type BLS vote/cert for
determining the genesis slot for Alpenglow.

#### Summary of Changes
Add the new vote type and enable tracking in certificate pool.
Follow up PRs will plug this in and wire the BLSSigverifier to accept
this vote type *only* during the migration period.