Add Migration tracker and update consumers (#537)
Split from #502 
Best reviewed commit by commit

#### Problem
We currently use the alpenglow feature flag to determine whether to
enable Alpenglow. This does not work with Carlgration.

#### Summary of Changes
Instead add a tracker `MigrationStatus` for use with all threads that
need to be Alpenglow aware. This PR only updates some of the consumers,
enabling alpenglow and updating the rest of the consumers will happen in
a future PR.

`MigrationStatus` will track:
- Whether alpenglow is enabled (poh is shutdown and block creation loop
is started)
- Our view of the genesis block - the ancestor of the super-oc block
- The cluster generated genesis certificate

In a future PR we will set our view of the genesis block based on
super-oc tracking, and enable alpenglow once it matches the certificate.

Additionally we need to store the genesis block in `Bank` and in a
future PR the snapshot. This is useful to distinguish the following
scenario:
- We startup and the feature flag is active.
- We do not know if this means we have already completed the migration,
or we are before/in the middle of migration.
- Instead we can check the root bank on startup to see if the genesis
block is present.

Finally we update replay to read `MigrationStatus`:
- On startup check the root bank to distinguish whether we are pre or
post migration.
- If the feature flag is activated during steady state, compute and
store the migration slot in `MigrationStatus`
- Use `MigrationStatus` to determine if we should `maybe_start_leader`,
send `VotorEvent::Block` or engage in any of the TowerBFT consensus
protocols.

PohService similarly is updated to use `MigrationStatus` to determine
when to shutdown and enable the block creation loop.