migration: do not set root for slots in migrationary period (#557)
Split from #502 

#### Problem
During the migrationary period it is not safe to set root. We might end
up rooting past the genesis block and clear it.

#### Summary of Changes
Only set root for slots < migration slot. We don't have to reenable this
after genesis, because at that point votor will take over setting roots
and this pathway in replay stage won't even be called.