migration: VoM for leader banks (#565)
Split from #502 

#### Problem
When we are in the migrationary period we don't want to include user txs
in blocks.

#### Summary of Changes
Change `maybe_start_leader` to make leader banks as VoM. This value is
used by banking stage to only include vote txs in packing. We do not
need to undo this after genesis as replay stage is no longer responsible
for starting leader - block creation loop will create leader banks as
normal.

NOTE: this is best effort, e.g. a malicious guy could comment this out.
Validation will be handled in #566