migration: go dark on OC RPC during migrationary period (#558)
Split from #502 

#### Problem
Although migrationary period blocks will not contain user transactions,
we would like to not report confirmation during this time as it will be
confusing to users - and we intend to reuse the same slot #s post
migration

#### Summary of Changes
For any slot > migration slot, do not report OC notifications. We don't
need to reenable this post genesis as Votor doesn't have a notion of OC
and instead we set this confirmation level via finalized slots.

No more TowerBFT votes so this thread in CIVL will just idle until it's
deleted in 4.1.