block_creation_loop: startup separately from PohService (#574)
#### Problem
Want to upstream the block creation loop, last bit of cleanup.
Currently it's a bit messy because it's started via PohService and poh
can't depend on core so it's smuggled in as a closure.
Eventually we will also delete PohService.

#### Summary of Changes
Separate it out into its own thread, it waits until PohService is
shutdown before starting block building.